+++ .storybook/preview.tsx
@@ -1,9 +1,18 @@
+import isPropValid from '@emotion/is-prop-valid'
 import type { Preview } from '@storybook/react'
 import React from 'react'
-import { ThemeProvider } from 'styled-components'
+import { StyleSheetManager, ThemeProvider } from 'styled-components'
 
 import { GlobalStyles } from '../src/globalStyles'
 import theme from '../src/theme'
+
+const shouldForwardProp = (propName: string, target: any) => {
+	return typeof target === 'string' ? isPropValid(propName) : true
+}
+
+const customShouldForwardProp = (prop: string, target: any): boolean => {
+	return !prop.startsWith('$') && shouldForwardProp(prop, target)
+}
 
 const preview: Preview = {
 	parameters: {
@@ -26,10 +35,12 @@ const preview: Preview = {
 	decorators: [
 		(Story) => (
 			<React.StrictMode>
-				<ThemeProvider theme={theme}>
-					<GlobalStyles />
-					<Story />
-				</ThemeProvider>
+				<StyleSheetManager shouldForwardProp={customShouldForwardProp}>
+					<ThemeProvider theme={theme}>
+						<GlobalStyles />
+						<Story />
+					</ThemeProvider>
+				</StyleSheetManager>
 			</React.StrictMode>
 		)
 	],