+++ src/components/Icon/IconStyles.tsx
@@ -1,43 +1,32 @@
 import SVG from 'react-inlinesvg'
-import styled, {
-	css,
-	DefaultTheme,
-	FlattenInterpolation,
-	FlattenSimpleInterpolation,
-	ThemedStyledProps
-} from 'styled-components'
+import styled, { css, RuleSet } from 'styled-components'
 
-import { ColorAndTheme, getTextColor } from '../../tokens/helpers'
+import { getTextColor } from '../../tokens/helpers'
 import { Color, ColorTheme } from '../../types/Color'
 import { CSSValue } from '../../types/CSSValue'
 import { PhoenixIcons } from '../../types/PhoenixIcons'
 import { getSpacingCssValue, Spacing } from '../../types/Spacing'
 import { isPhoenixIconColored } from '../../utils/icons'
+import { MarginProps } from '../common/Spacing/MarginProps'
 import { marginCss } from '../common/Spacing/SpacingStyles'
 
 interface StyledIconProps {
-	$size: Spacing | CSSValue
-	$color?: Color
+	size: Spacing | CSSValue
+	color?: Color
 	icon: PhoenixIcons
 }
 
-export const StyledIconContainer = styled.span<StyledIconProps>`
+export const StyledIconContainer = styled.span<StyledIconProps & MarginProps>`
 	display: inline-flex;
 
-	${({ $size, theme }): FlattenSimpleInterpolation => css`
-		width: ${getSpacingCssValue(theme, $size)};
-		height: ${getSpacingCssValue(theme, $size)};
+	${({ size, theme }): RuleSet => css`
+		width: ${getSpacingCssValue(theme, size)};
+		height: ${getSpacingCssValue(theme, size)};
 	`}
 
 	${marginCss}
 
-	${({
-		theme,
-		$color,
-		icon
-	}):
-		| FlattenInterpolation<ThemedStyledProps<ColorAndTheme, DefaultTheme>>
-		| undefined => {
+	${({ theme, color, icon }): RuleSet | undefined => {
 		if (isPhoenixIconColored(icon)) {
 			const colorTheme = icon.split('-').slice(-1)[0] as ColorTheme
 
@@ -50,7 +39,7 @@ export const StyledIconContainer = styled.span<StyledIconProps>`
 				}
 			`
 		}
-		if ($color) {
+		if (color) {
 			return css`
 				path {
 					fill: ${getTextColor()};