+++ src/components/Menu/MenuItem.tsx
@@ -19,7 +19,7 @@ export interface MenuItemProps
 		GenericComponentProps {
 	target?: string
 	href?: string
-	onClick?: MouseEventHandler<HTMLAnchorElement>
+	onClick?: MouseEventHandler<HTMLButtonElement | HTMLAnchorElement>
 	icon?: IconType | PhoenixIconsColored
 	colorTheme?: ColorTheme
 	disabled?: boolean
@@ -37,7 +37,7 @@ export const MenuItem: React.FC<PropsWithChildren<MenuItemProps>> = ({
 	...restProps
 }) => {
 	const getClickHandler = ():
-		| MouseEventHandler<HTMLAnchorElement>
+		| MouseEventHandler<HTMLButtonElement | HTMLAnchorElement>
 		| undefined => (disabled ? undefined : onClick)
 
 	const commonProps = {